refactor(arch): FSD cleanup — PR #491 Suggestion 수정
- S1: _HowItWorksSkeleton dead code 삭제
- S2: ContactForm.test.tsx jest.mock → barrel 경로
- S3: CrossLinkCards app 레이어 4곳 deep import → barrel
